+++ packages/core/src/components/tree-view/tree-view.interfaces.ts
@@ -1,6 +1,6 @@
-export interface TreeViewItem<T> {
-  /** The reference of type T */
-  reference: T;
+export interface TreeViewItem {
+  /** The id of the item */
+  id: string;
   /** The label of the item */
   label: string;
   /** The optional href of the item (creates a link) */
@@ -8,11 +8,15 @@ export interface TreeViewItem<T> {
   /** Indicates whether the item has children */
   hasItems: boolean;
   /** The array of child items */
-  items?: TreeViewItem<T>[];
+  items?: TreeViewItem[];
   /** Indicates whether the node is open and child items are shown */
   open?: boolean;
   /** Indicates the node is loading child items */
   loading?: boolean;
+  /** Indicates the node is active, only one item should be active */
+  active?: boolean;
+  /** Indicates the node is selected, multiple items can be selected */
+  selected?: boolean;
   /** An optional array of icons */
   icons?: TreeViewItemIcon[];
 }
@@ -22,4 +26,11 @@ export interface TreeViewItemIcon {
   icon: string;
   /** The label for the icon */
   label: string;
+}
+
+export interface TreeViewPointerEvent {
+  /** The path to the clicked item */
+  path: TreeViewItem[],
+  /** The original pointer event */
+  originalEvent: MouseEvent
 }