+++ packages/react/src/components/tree-view/tree-view.template.tsx
@@ -9,7 +9,7 @@ export function treeViewTemplate(
     onOpenItem,
     onCloseItem,
     onClickItem
-  }: TreeView<string>
+  }: TreeView
 ) {
   return (
     <DsoTreeView